+++ src/Bridge/Cloudflare/CHANGELOG.md
@@ -1,6 +1,15 @@
 CHANGELOG
 =========
 
+0.11
+----
+
+ * Introduce a `StoreFactory`
+ * [BC BREAK] Add support for `ScopingHttpClient` in `Store`
+ * [BC BREAK] The `accountId` parameter for `Store` has been removed
+ * [BC BREAK] The `apiKey` parameter for `Store` has been removed
+ * [BC BREAK] The `endpointUrl` parameter for `Store` has been removed
+
 0.4
 ---